Basic Setup Complete

+++ engine/config/vec2.h
@@ -4,13 +4,13 @@
 class vec2
 {
 public:
-	double x = 0;
-	double y = 0;
+	float x = 0;
+	float y = 0;
 
 	vec2() {};
-	vec2(unsigned int xIn, unsigned int yIn) : x(double(xIn)), y(double(yIn)) {}
+	vec2(unsigned int xIn, unsigned int yIn) : x(float(xIn)), y(float(yIn)) {}
 	vec2(int xIn, int yIn) : x(xIn), y(yIn) {}
-	vec2(double xIn, double yIn) : x(xIn), y(yIn) {}
+	vec2(float xIn, float yIn) : x(xIn), y(yIn) {}
 
 	inline vec2 operator + (const vec2& vec) const
 	{
@@ -22,12 +22,12 @@ public:
 		return vec2(x - vec.x, y - vec.y);
 	}
 
-	inline vec2 operator / (double val) const
+	inline vec2 operator / (float val) const
 	{
 		return vec2(x / val, y / val);
 	}
 
-	inline vec2 operator * (double val) const
+	inline vec2 operator * (float val) const
 	{
 		return vec2(x * val, y * val);
 	}
@@ -60,36 +60,36 @@ public:
 		y -= vec.y;
 	}
 
-	inline void operator *= (double val)
+	inline void operator *= (float val)
 	{
 		x *= val;
 		y *= val;
 	}
 
-	inline void operator /= (double val)
+	inline void operator /= (float val)
 	{
 		x /= val;
 		y /= val;
 	}
 
-	inline double distSq(const vec2& vec) const
+	inline float distSq(const vec2& vec) const
 	{
 		return (x - vec.x) * (x - vec.x) + (y - vec.y) * (y - vec.y);
 	}
 
-	inline double dist(const vec2& vec) const
+	inline float dist(const vec2& vec) const
 	{
 		return sqrt(distSq(vec));
 	}
 
-	inline double length() const
+	inline float length() const
 	{
 		return sqrt(x * x + y * y);
 	}
 
 	inline vec2 normalize() const
 	{
-		double l = length();
+		float l = length();
 		return vec2(x / l, y / l);
 	}
